Update scripts to have nice logs
