+++ .changeset/short-weeks-give.md
@@ -1,0 +1,5 @@
+---
+"chatbot": patch
+---
+
+Add a logs format to all the modules scritpts
