+++ apps/chatbot/src/modules/logger.py
@@ -1,0 +1,19 @@
+import logging
+
+
+def get_logger(name: str) -> logging.Logger:
+    formatter = logging.Formatter(
+        fmt="%(asctime)s %(levelname)s [%(name)s] [%(funcName)s]: %(message)s"
+    )
+
+    handler = logging.StreamHandler()
+    handler.setFormatter(formatter)
+
+    logger = logging.getLogger(name)
+    logger.setLevel(logging.DEBUG)
+
+    if not logger.handlers:  # Prevent adding multiple handlers
+        logger.addHandler(handler)
+        logger.propagate = False
+
+    return logger
